# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ReviewBot AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface PRFetchInput {
  owner: string;
  repo: string;
  prNumber: number;
}

export interface PRFile {
  filename: string;
  /** @nullable */
  patch?: string | null;
  status: string;
  additions: number;
  deletions: number;
}

export interface PRData {
  title: string;
  author: string;
  branch: string;
  baseBranch: string;
  filesChanged: number;
  additions: number;
  deletions: number;
  owner: string;
  repo: string;
  prNumber: number;
  files: PRFile[];
}

export interface ReviewInput {
  files: PRFile[];
}

export type ReviewCommentSeverity = typeof ReviewCommentSeverity[keyof typeof ReviewCommentSeverity];


export const ReviewCommentSeverity = {
  critical: 'critical',
  warning: 'warning',
  suggestion: 'suggestion',
} as const;

export type ReviewCommentCategory = typeof ReviewCommentCategory[keyof typeof ReviewCommentCategory];


export const ReviewCommentCategory = {
  security: 'security',
  bug: 'bug',
  performance: 'performance',
  'code-quality': 'code-quality',
  'best-practice': 'best-practice',
} as const;

export interface ReviewComment {
  id: string;
  severity: ReviewCommentSeverity;
  category: ReviewCommentCategory;
  file: string;
  line: number;
  title: string;
  issue: string;
  fix: string;
  explanation: string;
  /** @nullable */
  cwe?: string | null;
}

export interface PostCommentsInput {
  owner: string;
  repo: string;
  prNumber: number;
  comments: ReviewComment[];
}

export interface PostCommentsResult {
  success: boolean;
  posted: number;
}