# lib/api-zod/src/generated/types/pRData.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ReviewBot AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PRFile } from './pRFile';

export interface PRData {
  title: string;
  author: string;
  branch: string;
  baseBranch: string;
  filesChanged: number;
  additions: number;
  deletions: number;
  owner: string;
  repo: string;
  prNumber: number;
  files: PRFile[];
}