# lib/api-zod/src/generated/types/pRFetchInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ReviewBot AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PRFetchInput {
  owner: string;
  repo: string;
  prNumber: number;
}