# lib/api-zod/src/generated/types/pRFile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ReviewBot AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PRFile {
  filename: string;
  /** @nullable */
  patch?: string | null;
  status: string;
  additions: number;
  deletions: number;
}