# lib/api-zod/src/generated/types/postCommentsInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ReviewBot AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReviewComment } from './reviewComment';

export interface PostCommentsInput {
  owner: string;
  repo: string;
  prNumber: number;
  comments: ReviewComment[];
}