# lib/api-zod/src/generated/types/reviewComment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ReviewBot AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReviewCommentCategory } from './reviewCommentCategory';
import type { ReviewCommentSeverity } from './reviewCommentSeverity';

export interface ReviewComment {
  id: string;
  severity: ReviewCommentSeverity;
  category: ReviewCommentCategory;
  file: string;
  line: number;
  title: string;
  issue: string;
  fix: string;
  explanation: string;
  /** @nullable */
  cwe?: string | null;
}