# lib/api-zod/src/generated/types/reviewInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ReviewBot AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PRFile } from './pRFile';

export interface ReviewInput {
  files: PRFile[];
}